refactor: address code review — SDK extraction, DRY, console split, plugin registry
Code review feedback from @EvanLeleux addressed:

SDK changes:
- Move conversation send/poll to sdk/entities/agents.py as agents.ask()
  (fixes stale follow-up responses by snapshotting messages before polling)
- Move fulltext search to sdk/entities/search.py as search.fulltext()
- Merge evidence into risks.get(key, evidence=True) instead of separate method

Handler refactors:
- handlers/agent.py: thin wrappers calling sdk.agents.ask(), no inline polling
- handlers/find.py: thin wrapper calling sdk.search.fulltext()
- handlers/engagement.py: removed group, moved create-customer under add group
  as 'guard add customer'. Removed vault creation (Marcus can't do this).
- handlers/run.py: removed hardcoded TOOL_ALIASES, capabilities resolved live
  from /capabilities/ API. resolve_target uses sdk.search.fulltext().
- runners/local.py: plugin/registry pattern replaces giant elif block

Console split (2141 lines → 8 modules):
- console.py: core loop, dispatch, banner (~400 lines)
- commands/context.py: set/unset/show/use/options/execute/back
- commands/accounts.py: accounts/switch/home
- commands/search.py: search/find/assets/risks/jobs/info (calls SDK)
- commands/tools.py: run/status/download/install/capabilities
- commands/marcus.py: ask/marcus/read/ingest/do (calls sdk.agents.ask)
- commands/reporting.py: evidence/report
- renderer.py: render_results, render_evidence, render_entity_detail

Other:
- Marcus conversation mode: commands prefixed with / (e.g., /back, /new)
- README: restored auth/signup sections, added console docs
- Created docs/console.md for LLM reference
- Deleted design spec doc
- marcus read docstring notes file_read tool availability
