+++ praetorian_cli/ui/console/commands/__init__.py
@@ -1,0 +1,6 @@
+from praetorian_cli.ui.console.commands.context import ContextCommands
+from praetorian_cli.ui.console.commands.accounts import AccountCommands
+from praetorian_cli.ui.console.commands.search import SearchCommands
+from praetorian_cli.ui.console.commands.tools import ToolCommands
+from praetorian_cli.ui.console.commands.marcus import MarcusCommands
+from praetorian_cli.ui.console.commands.reporting import ReportingCommands
